View version changes now remove trees
While working on an update to Files, I caused a reindex on a files
collection with ~70k files. This reindex exposed a bug in Nebari, which
thankfully was actually much harder to reproduce than I had hoped --
trying to recreate the database didn't work, as it was a test database
that had evolved over time.

In the process of trying to figure out how to recreate a similar setup
in a test in Nebari to try to reproduce the bug, I realized that
remapping after a view version change really shouldn't try to do a lazy
update at all. The view query system will not return data from an old
version, and worse, if the map function's implementation changes such
that certain keys will not be emitted ever again, there will be phantom
entries in the view entries table that will forever cause wasted
processing cycles because they were never deleted.

By deleting the trees, we improve everything about the remap behavior --
disk bloat is reduced, and due to the tree being completely clean, the
insert speed should be better as well.