Removed most String error variants
The main motivation of this change was to make it easier to add more
crates to the bonsaidb ecosystem while keeping error handling simple.
This provides a mechanism to allow users to bubble all bonsaidb-related
errors up as bonsaidb::core::Error if they wish, as each of the error
types can convert to it. However, those conversions remove the original
error type and only preserves its Display representation. This is due to
Error needing to support serialization and deserialization.

For users who want to match on the error type, they can use the more
specific error type. For those looking for easy error handling, that now
works too.