Rewrote view and document storage
Both #76 and #225 ended up being heavily intertwined. This is not
yet in its final form, but it's complete enough that unit tests are
passing (aside from backwards compatibility ones).

Document Storage:

Documents are no longer serialized in a wrapper document type. Instead,
the documents tree is now a versioned tree with an embedded index that
stores the document's hash. The Revision's id is now the versioned
tree's sequence_id.

This means that instead of simply pulling a document out of the database
and deserializing it, we must pull the value and index out for a key and
combine it with the key to create our document.

The other major change is introduced by the constraints of working
within Nebari's modification system. Because we don't have access to the
index for a key we're about to set, most of the logic for creating the
OperationResult has been moved outside of the CompareSwap operation.

View Storage:

Views have been refactored to store the reduced value in Nebari through
use of an embedded index. Instead of storing the entire ViewEntry
structure in the view, we now only store the serialized
`Vec<Entrymapping>`. The major change here is that Nebari will now
reduce the stored index via the new `ViewIndexer`. The changes haven't
been made to reduce/reduce_grouped yet to use Nebari's native reduce
function -- but that is the inspiration for these changes.

When retrieving a view entry, we reconstruct the ViewEntry using the
stored index to maintain compatibility with the existing code that
worked with the ViewEntry structure.

There are a lot of remaining tasks:

- Update reduce/reduce_grouped() to use Nebari's built-in reduction.
- Remove the invalidated_entries map and make the view mapper sequence
  based.
- Embed the DocumentMap tree in the ViewEntries tree by creating a
  custom Root.
- Once all the above are done, when the view indexer is running outside
  of a transaction (lazy views), the view can be persisted without fsync
  and be 100% safe to use due to the append-only file format.