Refactored view mapping to be sequence based
This commit removes the invalidated entries tree, and uses the sequence
index of the documents tree to drive the indexing. The mapping operation
is batched and performed in such a way that if new data is added to the
documents tree while the operation is being performed, the indexing is
performed using the sequence data at the time of the mapping job being
kicked off.

This guarantee allows us to track what the latest indexed sequence ID in
the ViewEntries embedded index. The start of the map job begins from the
ViewEntries tree's latest sequence id + 1.