Fixed reindexing removed documents
This was a weird one to debug, as it only showed up on the
simultaneous-connections test. Yet, the bug was unrelated to
multiprocessing.

Eager views are meant to always be up-to-date. This contract was broken
when multiprocessing was involved, because there was a logic bug: the
index being returned from TransactionTree::remove is the existing index,
which means its sequence id is of the removed sequence, not the newly
writeten sequence (document entries aren't actually removed, for history
preservation).

The fix is to retrieve the new sequence value and map it instead. This
ensures we're actually mapping the deleted version of the entry.

The reason this didn't cause issues outside of multithreading is most
tests are written without specifying an access policy, which means all
the queries are AccessPolicy::UpdateBefore. This meant that the
preparation for queries would still index it, as it wasn't actually
up-to-date.