Embedded document map
This change turns ViewEntries into a new Root implementor for Nebari
that stores the view entries in one B+Tree, and stores the document map
in another B+Tree.

This pull request does not yet add the ability to query from the
document map. Once that is implemented, I can remove the external
document map tree which will conclude the final format changes.