Removed the document map
This removes the document_map tree, and stores it inline in a new custom
Nebari Root. This custom tree supports querying what keys a document id
emitted as well as what mappings were emitted for any given key.

This branch also contains several other changes:

The integrity scanner can spawn a mapping job, and that mapping job must
use transactions if the view is eager. This set of changes addressed
that, but it also lumped in with a refactor to change from
easy_parallel to rayon.

While rayon is a heaver dependency, I was noticing a *lot* of traffic on
profiles for spinning up new threads. Rayon uses a persistent thread
pool for work, and by embracing it here, we can start using it in other
locations as well.