Testing reuse_address from Fabruic
This commit hard-codes reuse_address in the QUIC layer. This isn't ideal
-- we need a way to provide this as a configuration. However, I wanted
to make sure the fix that's in a PR on Fabruic worked to fix this test.

That's when I discovered the same error eating behavior that the
websockets demonstrated. This probably needs to be fixed on wasm too.