Added Key::CAN_OWN_BYTES and other ByteCow-changes
For the deprecated functionality, I've reverted it back to only
supporting byte slices, and removed its ability to borrow. This is a
tradeoff to the hidden reallocations happening when using it, and it
greatly simplifies the code.

The other main change is to allow Option<T> and Result<T,E> to support
receiving an owned Vec<u8> when decoding, but allow it to be opt-in.
This is what CAN_OWN_BYTES controls. This allows Option<Vec<u8>> to
avoid an extra reallocation by using a memcpy to remove the front byte
rather than reallocating a new Vec from the subslice.

The other changes are just small changes to either fix clippy warnings
or fit my preferred approach.