Schema summaries are now available
Closes #278.

In a quest to allow clients that do not have the Rust type definitions
available to be able to query and interact with a BonsaiDb server, the
list_available_schemas functions have been updated to return a
`Vec<SchemaSummary>` instead of just the `SchemaName`s.

This change should allow generic database clients/admins to be built.