Composite key encoding now can escape nulls
After #254, null bytes no longer need to be either allow/deny. Instead,
we can now escape which is the default behavior. This will cause extra
allocations when null bytes are encountered during the encoding process,
but the correct sort behavior is preserved.

Allow/deny is still an option for situations where users do not want the
extra overhead that encoding can incur.