Optimize Key for basic enums
Enums that have no contained data are now encoded directly as their
representation without the overhead of the composite key encoder.

The error type is still CompositeKeyError, because we need an error type
that we can construct in the event of no enum matches. Maybe this should
be its own error type in the future.