Removed case insensitivity of database names
I'm not really sure what I was thinking when I implemented this
originally. The motivation was thinking about handling case-insensitive
filesystems. I knew that this wasn't a good approach for non-ascii, but
it didn't really hit me until tonight that there's no good approach --
the case insensitivity implementations of any given filesystem is what
we're at the mercy of, and there's no guarantee every filesystem follows
the same rules.

Normalizing case is one approach, but that still isn't a perfect match.
Regardless, the way it currently was implemented was completely broken.
This allows different cases in database names, but on case-insensitive
filesystems, using different cases will result in broken behavior.