Adding into_owned/Owned for Key
Refs #284

These changes add the ability for key types such as `Cow<'a, str>` to
convert to `Cow<'static, str>`, which will allow View/ViewSchema to
differentiate between the static/owned representation.

This is the least complicated change to #284. I attempted those changes
as well, but they're very involved, and this chunk of changes is
separate from those changes.