Added ViewSchema::MappedKey
Closes #284

This commit adds an associateed type to ViewSchema which is used in map
and reduce. This type uses a generic associated lifetime, which allows
for the map/reduce functions to utilize borrowed data.

The extra associated type is an unfortunate side effect of how lifetime
extension with generic associated types currently works -- see #287 for
where the exploration ended when trying to put the associated type on
`Key` instead.

Thanks to @asonix for their help in various ideas while experimenting on
how to make this work!