Split ViewSchema/MapReduce, add derive(ViewSchema)
This commit aims at reducing the complexity introduced in #284. By
moving map/reduce to their own trait, it made it easier to provide a
derive macro for ViewSchema.

The derive macro is able to substitute in the correct values when no
lifetimes are involved, and when the 'doc associated lifetime is used,
everything just works as intended.

This also has another nice side effect of simplifying where the magic
happens for CollectionDocument-based map/reduce. Now all the shared
ViewSchema functionality lives on the shared trait, and
CollectionMapReduce is a simplified version of MapReduce when working
with SerializedCollections.

Overall, this is a big win. The only major thing missing from this is UI
tests for the new macro attribute.