Introduced ViewSchema::update_policy
Also added documentation for deriving ViewSchema.

This change was inspired by me writing the documentation for deriving.
The fact that `unique()` and `lazy()` were separate allowed nonsensical
configurations that `bonsaidb-local` just ignored -- for example,
returning true from both lazy and unique seems invalid, but
bonsaidb-local just ignores lazy when unique is true.

Switching to an enum to represent the update policy makes everything
much clearer, and the documentation is easier to write and understand.