Implemented client-side timeouts
Refs #296, #297

This commit adds timeout settings for connections and requests. These
timeouts are separate, meaning that if the initial request takes X
seconds to connect, the request timeout will be timed in addition to
those X seconds.

This should be unit-tested, and the wasm client changes haven't been
tested at all beyond ensuring compilation.